requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
dj-rest-auth==2.1.3
Django==3.1.5
django-allauth==0.44.0
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-grappelli==2.14.3
djangorestframework==3.12.2
idna==2.10
isort==5.7.0
oauthlib==3.1.0
Pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
